requirements.txt
bentoml==1.0.0a4
dagshub==0.1.8
deepchecks==0.6.1
hydra_core==1.3.2
hyperopt==0.2.7
joblib==1.1.0
mlflow==1.25.1
numpy==1.22.4
omegaconf==2.1.2
pandas==1.4.2
pandera==0.13.4
patsy==0.5.2
pydantic==1.9.1
pytest_steps==1.8.0
reuests==2.28.0
scikit_learn==1.2.1
streamlit==1.10.0
xgboost==1.6.1